Move use_mcp_tool and access_mcp_resource to tool files
